Replaced tuned hard-included repo with a git submodule
Until now tuned repo was hard included in the NTO repo.
This design could lead to developers mistakenly contribute to assets/tuned.
Instead of https://github.com/redhat-performance/tuned, as it is hard to realize it is an external repo.
Therefore, we chose to change this design by including the tuned repo as a git submodule, preventing mistakes as I described and better informing tuned is an external repo.

Signed-off-by: Ronny Baturov <rbaturov@redhat.com>