pao:status: replace methods with functions
Decouple the `status.go` file from the controller.
Move it into separate package named (obviously) status.
Convert all methods into functions which also accepts `context` and `client`.

No changes in functionality are expected.

The motivation for this change is to be able to use the same
utility functions on both hypershift and non-hypershift clusters.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>