[release-4.15] [manual] OCPBUGS-30507: Add performance real time tuned template (#984) (#1025)
* Backport: Add performance real time tuned template (#954)

* Add performance real time tuned template

Kernel parameters that are not supported on RT kernel systems are being applied, causing errors to be logged to the tuned daemon, resulting in a degraded profile state.
Therefore, I added the openshift-node-performance-rt profile that would be included if an RT kernel is detected, thereby dropping the unsupported kernel parameters before they are applied.

* Added a test that make sure the tuned profile is not degraded

This could be great to have a generic test that make sure that after
performance profile is applied the tuned profile is not degraded.
This would prevent future issues like OCPBUGS-23167.

* render-sync update

Updated render-sync to include artifacts that are needed for the e2e tests.
Committing the rendered items to this commit.



* NO-JIRA: Update tuned profile degraded test (#1005)

* Added isVM func to util



* Update tuned profile degraded test

Updated the tuned degradation test to fail only when the tuned profile was found degraded
on BM host.
In case of tuned profile found degraded on a VM, only a warning would be reported in the logs.
The reason for that is that the fact CI is using VMs - and some
configurations like certain kernel args can't be applied and therby lead to
the tuned profile being degraded.



---------



---------

Signed-off-by: rbaturov <rbaturov@redhat.com>
Signed-off-by: Ronny Baturov <rbaturov@redhat.com>