CNF-6517: [Part 1]  Hypershift PAO adoption  (#1032)
* pao:resources: replace methods with functions

Decouple the `resources.go` file from the controller.
Move it into separate package named (obviously) resources.
Convert all methods into functions which also accepts `context` and `client`.

No changes in functionality are expected.

The motivation for this change is to be able to use the same
utility functions on both hypershift and non-hypershift clusters.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* pao:controller: replace methods with functions

decouple `getContainerRuntimeName` and `getMachineConfigPoolByProfile` from the receiver.
This is needed so we could reuse them outside the receiver scope.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* pao:controller: pass profile name instead of struct

Where possible, change the functions signature to accepts profile name
instead the whole profile struct.

This makes the functions:
1. more explicit about the input needed.
2. useable for hypershift cluster (which uses `configMaps` instead of `performanceProfile`

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* pao:controller: `getContainerRuntimeName` profile -> labels

Change the function to accepts labels instead of performanceProfile.
This way makes the function reuseable with hypershift.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* pao:status: remove unneeded receiver from functions

The functions are not using the receiver, let\'s remove it.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* pao:controller: `instance.Name` -> `instance.GetName()`

`instance.GetName()` uses the underlying object interface which is more generic.
in hypershift the instance is a ConfigMap and not PerformanceProfile so
we can use `instance.GetName()` for both cases.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* pao:status: replace methods with functions

Decouple the `status.go` file from the controller.
Move it into separate package named (obviously) status.
Convert all methods into functions which also accepts `context` and `client`.

No changes in functionality are expected.

The motivation for this change is to be able to use the same
utility functions on both hypershift and non-hypershift clusters.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* pao:controller: convert `getInfraPartitioningMode` to function

On hypershift the `getInfraPartitioningMode` should be running against the hosted cluster API, hence it needed
a different client. let's pass the client from the caller to make this function compatible for hypershift.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

---------

Signed-off-by: Talor Itzhak <titzhak@redhat.com>