CNF-6517: [Part 2] Hypershift PAO adoption (#1037)
* pao:components: components handler interface

Introduce an interface that allows customize handling
for the different components owned by the PerformanceProfile controller.

This way we can have a different implementation for hypershift and non-hypershift platform.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* pao:components: implement handler interface

Implement the handler interface when the controller is running on non-hypershift platform.

`Apply`  function is very similar to `applyComponents`:
https://github.com/openshift/cluster-node-tuning-operator/blob/master/pkg/performanceprofile/controller/performanceprofile_controller.go#L542

`Delete` function is the old `deleteComponents` function:
https://github.com/openshift/cluster-node-tuning-operator/blob/master/pkg/performanceprofile/controller/performanceprofile_controller.go#L619

`Exist` is the old `isComponentsExists` function:
https://github.com/openshift/cluster-node-tuning-operator/blob/master/pkg/performanceprofile/controller/performanceprofile_controller.go#L641

The actual wiring and initialization of this new handler in the reconciliation loop,
 will be done on later PR.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* pao:status Writer interface

Introduce an interface that allows customize writing of the status
for the the PerformanceProfile object.

This way we can have a different implementation for hypershift and non-hypershift platform.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* pao:resources: move `getMachineConfigPoolByProfile` to `resources`

This is needed so we could use it outside the controller package in
future implementations.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* pao:status: implemet Writer interface

Implement the Writer interface for non-hypershift platform.

Update function takes care of writing the status to the
PerformanceProfile object.
Similar to: https://github.com/openshift/cluster-node-tuning-operator/blob/master/pkg/performanceprofile/controller/status.go#L34

UpdateOwnedConditions function retrieves the owned objects conditions and
based on this information, updates the PerformanceProfile stauts.

The actual wiring of this new writer will be done on later commit.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* pao:controller: wire Writer and Handler interfaces

* Wire the new interfaces into the reconciliation loop.

* Remove all duplicated functions (which we copied into the Handler and
  Writer intefaces).

* Adapt the unit tests for the new implementation.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* pao:controller: generalize functions to accept client.object

On hypershift we would accept ConfigMap and not the PerformanceProfile.
Having this change allows handling both cases.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

---------

Signed-off-by: Talor Itzhak <titzhak@redhat.com>