fix extra-reboot on upgrade with paused mcp worker (#1053)
This PR fixes an extra-reboot when upgrading OpenShift with paused
MCP "worker", for example during EUS-to-EUS upgrade. The extra-reboot
occurred because PerformanceProfile controller was reconciling against
rendered MC appearing in the MCP status. While this MC reflects current
MCP conditions, it does not reflect the latest planned state when the
MCP is paused.
Unpausing the MCP lead to one reboot when applying the target MC, and
one additional reboot after the performance profile was reconciled
against it.
This change eliminates this additional reboot because PerformanceProfile
is now reconciling against the latest planned MC before the MCP is
unpaused

Signed-off-by: Vitaly Grinberg <v.g@redhat.com>
Co-authored-by: Vitaly Grinberg <v.g@redhat.com>