CNF-11817: e2e: control and data planes clients (#1004)
* pao:hypershift: `ControlPlaneClient`

On hypershift, there is a full separation between the control and data plane.
This means there's a separate client for each.

This commit implements the control plane client.
In order to preserve compatibility between HCP and OCP,
the client encodes/decodes the objects in/from a `ConfigMap` automatically.

This allows the user to perform similar CRUD operations without having the need
to distinguish between HCP and OCP platforms.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* e2e:data planes client

On Hypershift we should be able to access both the management and hosted clusters.
For that, we should now have two clients, each communicates with a different cluster.

ControlPlaneClient - defines the API client to run CRUD operations on the control plane that will be used for testing.

DataPlaneClient - defines the API client to run CRUD operations on the data plane that will be used for testing.

On non-hypershift both client would point to the same the cluster.

Also added Ginkgo assertions that prevent the user from using the `DataPlaneClient` client for the wrong object.
For example using the `DataPlaneClient` for querying the `PerformanceProfile` object
is completley fine on non-hypershift cluster, but will fail badly on hypershift (since `PerformanceProfile`
objects, are not presented on the hosted clusters.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* e2e: hypershift utils

Added functions for building the clients for hypershift while running on CI environment

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

---------

Signed-off-by: Talor Itzhak <titzhak@redhat.com>