tuned: controller: profiles extract/repack
rework profilesExtract and helper functions
to make more testable, enable round-tripping,
and add helpers which will be used in the deferred
update functionality.

Read extracted profiles on startup and reconstruct
the node state, to be used to implement the
deferred updates feature.
Please note all the operations are non-destructive
read-only so pretty safe.

Signed-off-by: Francesco Romani <fromani@redhat.com>