tuned: set the deferred flag
Now that all the pieces are in place, we can set
and use the deferred update flag to implement
the desired behavior.

Note we now need to trigger an event when the recommend file
is written, and always do that even if does not change.
This is needed to react properly when a deferred
update is un-deferred by editing the tuned object.

Without a trigger in this state the k8s object
status wouldn't be correctly updated.

Signed-off-by: Francesco Romani <fromani@redhat.com>