OCPBUGS-28647: tuned: operand: deferred updates prepare: cleanups from #1019 (#1073)
* Makefile: enable environment variables override

use the `?=` construct to let the Makefile variable
be easily overridden from the shell environment.

Signed-off-by: Francesco Romani <fromani@redhat.com>

* operator: unexport TunedProfiles

TunedProfiles -> tunedProfiles
trivial rename. The function was not used elsewhere anyway.

Signed-off-by: Francesco Romani <fromani@redhat.com>

* operator: profilecalculator: pack return values in a struct

Simplify the interface packing the many return values in
a single struct, to enable further extensibility.
No intended changes in behavior.

Signed-off-by: Francesco Romani <fromani@redhat.com>

* tuned: controller: simplify flow

Gather the node name once and use it in the lifecycle of
the controller - is not supposed to change anyway.
Bubble up the client creation, so we don't need to
keep around the kubeconfig.

Signed-off-by: Francesco Romani <fromani@redhat.com>

* tuned: changeSyncerProfileStatus align to left

Align to left for readability with no intended changes in behavior.

Signed-off-by: Francesco Romani <fromani@redhat.com>

---------

Signed-off-by: Francesco Romani <fromani@redhat.com>