CNF-11559: [Part 3/3] Hypershift PAO adoption (#1057)
* pao:hypershift: do not add mco scheme

On hypershift we don't have mco, trying to add it will cause for exit
with an error

* pao:hypershift: implement the components handler interface

This is the core of the reconciliation loop, which creates/updates/deletes the relevant components.
Similar to what we already have on OCP.

There's a gap about how to retrieve the `containerruntimeconfig` object
which is needed by the controller and is not accessible in the hosted
control plane namespace. This will be handle on a separate PR.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* pao:hypershift: stub implementation for status writer interface

status reporting is not supported on hypershift yet.
we provide a fake implementation in order to conforms the interface.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* pao:hypershift: check if in hypershift first

Checking the feature gate first will result with an error on hypershift since it's not yet supported.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* pao:hypershift: add management client

Adding a management client to be used on a hypershift platform.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* pao:hypershift: configure controller

Add the necessary configuration for the controller, so it would be able to run on a hypershift platform.

1. It needs to configure predicates for ConfigMaps instead of actual objects.
2. It needs to setup different clients for the management cluster and hosted cluster.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* pao:hypershift: generelize reconciliation loop

Modify the reconciliation flow so it could fit for both HCP and OCP.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* e2e:config: adapt suite for hypershift

Adding the relevant code for configuring performanceProfile over hypershift

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* hypershift:client: do not assign to nil map

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* hypershift:client: return error when configMap has invalid data state

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* pao:hypershift: move label validation to helper function

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* vendor: bump deps

Bump needed for handling hypershift APIs.
This is the output formatted and generated by the \`go mod tidy\`

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* vendor:featuregate: adopt feature gate api changes

The package that holds featuregate geatures has changed in the last vendor bump.
This commit uses the new paths.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* pao:log: small refining

- Move logs to V(4)
- Fix typos
- Remove redundant spacing
- Consistent HyperShift naming

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

---------

Signed-off-by: Talor Itzhak <titzhak@redhat.com>