e2e: modified client usage to controlplane and dataplane
Modified the functions and client calls in the 1_performance suite according to resource location. For instance, resources like the performance profile that reside in the management cluster will need to utilize the control plane client, whereas resources that reside on the hosted cluster will utilize the dataplane client.
Reminder: On OCP, both control plane and dataplane clients will be based on the same KUBECONFIG, so they will have the same effect.

Signed-off-by: Ronny Baturov <rbaturov@redhat.com>