hypeshift:client: fix Get function
There was a hidden (wrong) assumption that the `ConfigMap` name is equal to the
internal object's name which encapsulated in the `ConfigMap`.

To make sure we return the correct internal object,
we should list all the `ConfigMap`s in the HCP namespace,
decode the data and look for the correct object name.

Since the decoding operation is relatively expesive,
we added a label `ConfigMapEncapsulatedKindKey` that would
help with listing only `ConfigMap`s that contains the correct kind.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>