NO-JIRA:pao:hypershift: fix client implementation (#1094)
* hypershift: add unit-test for client Get() implementation

Turns out that the existing implementation of the `Get` call was buggy.
Adding a unit-test to make sure the issue was fixed.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* hypershift: rename `ObjIsEncapsulatedInConfigMap` -> `IsEncapsulatedInConfigMap`

We're passing the object as a parameter to the function,
so it stems from the function signature that it refers to the object and
we can remove the duplication in the function name (the `Obj` part).

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* pao:hypershift: fix Get() implementation

On hypershift everything is encapsulated in ConfigMap.
So if we're given the name of the encapsulated object,
trying to get the ConfigMap with same name is not enough,
because the internal object's name and the ConfigMap name
doesn't have to be equal.

Hence, the client have to `List()` all the ConfigMaps in the
hosted control plane namespace, extracts the data and
compares the names.

In case it was not found, it mimics an IsNotFound error
to be compatible with the error returned from the API server.

The implementation can be improved, by caching some of the information,
but before we come with "fancy-but-not-proved-to-be-efficent" solutions
let's evaluate the simplest solution first and iterate later if needed.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* main:scheme: add machineconfig scheme

The adjusments in the client implementation, requires
adding the scheme of machineconfig API

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* logs: add log when PerformanceProfile decoded successfully

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* pao:hypershift: check only for the existance of `RuntimeClass`

On hypershift we are not suppose to delete all the other objects
(i.e. MachineConfig/KubeletConfig/Tuned)

Those object has an ownerReference and will be removed by k8s garbage collector
when the PerformanceProfile gets deleted.

See: https://github.com/openshift/cluster-node-tuning-operator/blob/0a2f7b33d52a21053e3b140ed4035ff8581f5a0a/pkg/performanceprofile/controller/performanceprofile/hypershift/components/handler.go#L55-L60

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* hypershift:runtimeclass: pass mutated object

Previously we were passing the manifested object which doesn't
have a `resourceVersion` and would failed on `Update()`

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* predicates: reconcile if owned object has changed

We should also reconcile in case that one of the owned object has been tampered.
This is usually not expected because the user doesn't have an access to
the hosted cluster namespace.
Nevertheless, we should make sure to act in a case since
the behavior might be changing in the future.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

---------

Signed-off-by: Talor Itzhak <titzhak@redhat.com>