e2e: deferred: assertions to avoid unwanted changes
Add assertions to ensure deferred updates don't
affect unrelated nodes.
In the e2e tests we target one of the worker nodes,
so we check control-plane nodes are not affected
by the testing changes.

Checking different MCPs would be equivalent or better,
but this is the simplest way to check because
the worker/control-plane split is always available.

Signed-off-by: Francesco Romani <fromani@redhat.com>