tuned: fix propagation of the deferred flag
When we compute `TunedProfile`s, we should propagate
the deferred flag only if one of these object is
extracted from a `Tuned` object which had the deferred annotation.

The best option would be to set the deferred flag if any of the `Tuned`
objects relevant for a node has the deferred flag set.
However, this approximation makes the flow actually more predictable,
so we adopt it.

Signed-off-by: Francesco Romani <fromani@redhat.com>