NO-JIRA: hypershift:unitests: adjust controller unit-tests (#1105)
* hypershift:component: remove receiver from function

We want to reuse this functions in the unit tests
so we remove the receiver and export it.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* error: change to pointer in error message

Report the accurate type in the returned error.
We are expeting pointer of Type and and not Type itself.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* unittests:controller: adjust for hypershift

We want same unit tests to be running for hypershift,
so we convert the tests into table and added one entry for OCP and one for HCP.

As for now most of the tests are not supported/qualified to be running on HCP and
are skipped for now.

But it is better and safer to have a low coverage than nothing.
We'll work to support additional tests in future PRs

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

---------

Signed-off-by: Talor Itzhak <titzhak@redhat.com>