CNF-13394: Changed node inspector to have lazy initialization (#1093)
* Changed node inspector to have lazy initialization

To streamline the deployment of the node inspector, we have implemented a change where the ExecCommand will automatically deploy the node inspector if it is not already running. This approach offers several benefits:

* On-Demand Deployment: The node inspector is deployed only when needed.
* Users no longer need to explicitly deploy the inspector using the 0_config suite, allowing them to run specific specs or suites without verifying the node inspector's deployment status.

Moreover, this commit introduces some API simplification as we abstracted this process from the user.
* Lowercasing functions that are not intended to be called from outside the package.
* It is obvious that we would always use the Dataplane client, hence removed it from all function signatures.
* Introducing global variables for the namespace and node inspector, as these values remain constant.

Signed-off-by: Ronny Baturov <rbaturov@redhat.com>

* Changed node inspector to use same context

Signed-off-by: Ronny Baturov <rbaturov@redhat.com>

* Added warning logs and improved error handling

* Added warning logs for the Delete and Create functions.
* Improved error handling

Signed-off-by: Ronny Baturov <rbaturov@redhat.com>

* e2e: added node inspector cleanup on every suite

Currently we are cleaning up the node inspector in the following two ways:
1. Z_deconfig that runs in the end of all suites.
2. We added a trap command in the hack/run-test.sh script to execute the Z_deconfig suite once tests finished or interrupted.

However, when we import some of these suites to an external repo (e.g cnf-features-depoly) none of the cleaning up methods described above are applied.
Therefore we risk that the node inspecor will be left as a side effect after tests complete.
Adding the cleanup after in the end of each suite, will reduce the risk of such a scenario to occur.
If we chose to delete after each suite, we can safely remove the Z_deconfig and the trap command as they are redundant.

---------

Signed-off-by: Ronny Baturov <rbaturov@redhat.com>