CNF-13530: hypershift: support `containerruntimeconfig` (#1100)
* hypershift: managementClusterNamespaceName -> hostedControlPlaneNamespaceName

"Hosted control plane namespace" is the name
that is being used by hypershift engineers and documentation
to describe the namespace where the hosted cluster control plane's componenets reside.
let's align the naming in our codebase.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* hypershift: modify DecodeManifest function signature

There's no indication return from the function whether the passed object
has loaded or not.

Now the function will return another boolean value to indicate that.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* hypershift:containerruntimeconfig: add support for different container runtimes

NTO supports different runtimes for the high-performance runtime class:
ref: https://github.com/openshift/cluster-node-tuning-operator/pull/588
This commit adds this kind of support when NTO runs on hypershift platform.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* unit-test: do not skip the containerruntime tests

Now we support it so let's test it for improved coverage

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

---------

Signed-off-by: Talor Itzhak <titzhak@redhat.com>