NO-JIRA: hypershift: implements missing client calls needed for e2e tests  (#1107)
* hypershift: add `Update()` call

The `Update()` call knows how to handle updates of the internal object that is being passed to it.

It compatible on both HCP and non-HCP platforms."

Co-authored-by: Ronny Baturov <rbaturov@redhat.com>
Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* hypershift: add `List()` call

The `List()` call knows how to handle list operation of the internal objects that is being passed to it.

It's compatible on both HCP and non-HCP platforms.

Co-authored-by: Ronny Baturov <rbaturov@redhat.com>
Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* hypershift: provide data key to client calls

By passing the data key to the client calls we can save redundant type checking
and reduce duplicated code.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

* hypershift: reuse `GetObjectConfigMapDataKey` to implement `IsEncapsulatedInConfigMap`

Signed-off-by: Talor Itzhak <titzhak@redhat.com>

---------

Signed-off-by: Talor Itzhak <titzhak@redhat.com>
Co-authored-by: Ronny Baturov <rbaturov@redhat.com>