CNF-7123: Enable and adjust 1_performance suite on Hypershift (#1084)
* e2e: openshift label for unsupported tests on hypershift

This way we can filter out unsupported tests when running on hypershift.

* The API conversion nor the validation webhooks are supported.
* On hypershift we don't support multiple performance profile on a hosted cluster, as one performance profile is applied to all worker nodes (nodepool).
* Due the bullet above, there is no non RT worker nodes. hence, tests relying on nodes without tuning are not supported.

Signed-off-by: Ronny Baturov <rbaturov@redhat.com>

* e2e: modified GetWithRetry to execute by client provided

Let's enable the support for both dataplane and control plane clients utilize the GetWithRetry function.

Signed-off-by: Ronny Baturov <rbaturov@redhat.com>

* e2e: hypershift: added GetByClusterName in nodepool API

This function logic will be needed, hence for reusability reasons,
I moved it to the nodepool API.

Signed-off-by: Ronny Baturov <rbaturov@redhat.com>

* e2e: modified client usage to controlplane and dataplane

Modified the functions and client calls in the 1_performance suite according to resource location. For instance, resources like the performance profile that reside in the management cluster will need to utilize the control plane client, whereas resources that reside on the hosted cluster will utilize the dataplane client.
Reminder: On OCP, both control plane and dataplane clients will be based on the same KUBECONFIG, so they will have the same effect.

Signed-off-by: Ronny Baturov <rbaturov@redhat.com>

* e2e: changed client Patch command to use the new Update command

Since, performance profile CRD is not registered in hypershift env(and is encapsulated in configMap),
we can't Patch a the performance profile configMap like we did before:
    testclient.Client.Patch(context.TODO(), profile,
        client.RawPatch(
            types.JSONPatchType,
            []byte(fmt.Sprintf(`[{ "op": "replace", "path": "/spec", "value": %s }]`, spec)),
        ))
Therefore, we will use instead the new control plane client Update function

Signed-off-by: Ronny Baturov <rbaturov@redhat.com>

* e2e: added WaitForTuningUpdating and WaitForTuningUpdated

On OCP, we use MCP conditions to determine if the tuning has started to update in the cluster and to indicate when it has finished. On Hypershift, the equivalent method is to check the UpdateConfig condition of the hosted cluster nodepool API.

Signed-off-by: Ronny Baturov <rbaturov@redhat.com>

* e2e: pao: replaced mcp.WaitForCondition calls

Replaced mcp.WaitForCondition calls to utilize the newly combined functions (WaitForTuningUpdating, WaitForTuningUpdated) that will support hypershift and OCP.

Note: On OCP, these functions will wait only on the worker-cnf MCP, and not on the worker node as the code did before. Since there was no use of the non-performance worker node in the irqbalance test, I think it is safe to omit that.

Signed-off-by: Ronny Baturov <rbaturov@redhat.com>

* makefile: adding 1_performance for hypershift testing

Signed-off-by: Ronny Baturov <rbaturov@redhat.com>

* e2e: hypershift: add worker-cnf label

This will ensure we always have one of the worker nodes with the worker-cnf label that will be utilized for the tests.
Since all nodes are RT worker nodes, we can choose arbitrarily the first one.

Signed-off-by: Ronny Baturov <rbaturov@redhat.com>
Co-authored-by: Talor Itzhak <titzhak@redhat.com>

---------

Signed-off-by: Ronny Baturov <rbaturov@redhat.com>
Co-authored-by: Talor Itzhak <titzhak@redhat.com>