hypershift:client: embed status in `PerformanceProfile`
On hypershift, PerformanceProfile's status is reported by a separate ConfigMap,
but for testing purposes we want to have same API as standard OCP.
IOW, we want to be able to access the status from the PerformanceProfile itself as follow:
`profile.status`

In order to achieve that, when user performs Get/List operation for PerformanceProfile,
the ControlPlane client will look for the ConfigMap that holds
the status (if it exists), extracts the status from it and embed it in the PerformanceProfile.

NOTE:
At this point in time the `Get()` and `List()` calls are done only
during the tests.
So this extra work of merging 2 ConfigMaps into the same `PerformanceProfile`
is quite expensive but is done only on testing anyway.
In the future, if we'll have to call `Get()`/`List()` with the `PerformanceProfile`
as part of the production code, we'll probably evaluate this approach
again.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>