e2e: performance: skip test on env with insufficient cpus (#1111)
Some CIs run this test on an environment that has exactly 2 isolated
cpus, but that doesn't mean that these are completely free cpus as there
is a node base load (cpu consumption of burstable and besteffort
pods).The problem is that this test will always fail on such
environments while it's irrelevant to run.

Skip the test by adding a condition to verify whether the requested cpus
are greater than what the node can provide.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>