hypershift: client: add `Delete` call
The `Delete` call deletes the `ConfigMap` that holds an encoded data of the object that was given for deletion.
This is needed for keeping backward compatability with API we're using for the
testing.
ref: https://github.com/openshift/cluster-node-tuning-operator/blob/10bb4c6a1388c91ee4184945e6103c82608eef27/test/e2e/performanceprofile/functests/3_performance_status/status.go#L153

NOTE:
This flow is not being used in production code ATM.
In case it does in the future, we might reassess this implementation if needed.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>