client:dataplane: more granual access for objects
So far, we had objects that are accessible on the control plane (management cluster)
OR objects that are accessible on the data plane (hosted cluster).

Tuned is a special object which gets created on the control plane but then
mirrored into the data plane.
Sometime we want to read Tuned from the data plane, but we still don't want
to allow its creation on the data plane directly.

Hence, we provides a different checks for Read operation and Write operarion
as part of the data plane implementation.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>