nto: export the logic for unique tuned name
On hypershift, `Tuned` object is being mirrored to the hosted cluster.
NTO makes `Tuned` names unique in case that Tuned is duplicated across `NodePool`s,
for example, if one `ConfigMap` is referenced in multiple `NodePool`s.

On OCP we used to list `Tuned` under the status of the
`PerformanceProfile`.

In order to achieve same behavior for hypershift,
we're capturing the logic, that grants unique name to `Tuned` objects,
into a function and export it.

On later commit, we'll use it to capture the name under the status.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>