controller: change logic for checking if MixedCPUs enabled
Since that on Hypershift, accessing the internal `PerformanceProfile`
isn't trivial, on the outter call the controller only checks if MixedCPUs is enabled by feature gate.

Later on, deeper in the stack the controller crosses the values
from both FeatureGate and the `PerformanceProfile` itself to determine
if MixedCPUs feature is enabled for this specific system.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>