OCPBUGS-28647: tuned: operand: add support for deferred updates (#1019)
* util: add helpers to deal with the deferred annotation

Add helpers to be used in future commits.
Enable early introduction of e2e tests

Signed-off-by: Francesco Romani <fromani@redhat.com>

* status: report deferred status

add support to report the deferred status in
conditions.

Signed-off-by: Francesco Romani <fromani@redhat.com>

* tuned: controller: profiles extract/repack

rework profilesExtract and helper functions
to make more testable, enable round-tripping,
and add helpers which will be used in the deferred
update functionality.

Read extracted profiles on startup and reconstruct
the node state, to be used to implement the
deferred updates feature.
Please note all the operations are non-destructive
read-only so pretty safe.

Signed-off-by: Francesco Romani <fromani@redhat.com>

* tuned: add changeSyncerPostNodeRestart

The deferred updates feature need to do specific
actions when the node on which the tuned daemon runs is restarted (and only then).
Add hook point with specific code for that.

Signed-off-by: Francesco Romani <fromani@redhat.com>

* tuned: refactor status update

Rework how tuned controller updates status to make
room for the deferred update feature.

Signed-off-by: Francesco Romani <fromani@redhat.com>

* tuned: set the deferred flag

Now that all the pieces are in place, we can set
and use the deferred update flag to implement
the desired behavior.

Note we now need to trigger an event when the recommend file
is written, and always do that even if does not change.
This is needed to react properly when a deferred
update is un-deferred by editing the tuned object.

Without a trigger in this state the k8s object
status wouldn't be correctly updated.

Signed-off-by: Francesco Romani <fromani@redhat.com>

* tuned: fix propagation of the deferred flag

When we compute `TunedProfile`s, we should propagate
the deferred flag only if one of these object is
extracted from a `Tuned` object which had the deferred annotation.

The best option would be to set the deferred flag if any of the `Tuned`
objects relevant for a node has the deferred flag set.
However, this approximation makes the flow actually more predictable,
so we adopt it.

Signed-off-by: Francesco Romani <fromani@redhat.com>

* tuned: operand: review log verbosiness

tune down the log chattiness to reduce the log pressure.

Signed-off-by: Francesco Romani <fromani@redhat.com>

* tuned: operand: log restarts

Log reasons why a tuned reload or restart
is requested. We do this at V>=4, so there's
little risk of adding extra noise in the logs.

Signed-off-by: Francesco Romani <fromani@redhat.com>

* tuned: operand: unified restart and reload handling

Previously, we added special logic to handle TuneD reload
requests when deferred updates are received.
Extend the logic to TuneD restarts for the sake of clarity
and because difform treatment was causing bugs:
profile not correctly reverted after removal.

Signed-off-by: Francesco Romani <fromani@redhat.com>

* e2e: makefile: add targets to build suites

building suites ahead of time enables trivial catching of
compilation error and turns out handy to run subset of tests,
in addition to the usual ginkgo filtering capabilities.

Signed-off-by: Francesco Romani <fromani@redhat.com>

* e2e: deferred: add test cases

add test coverage for the `deferred updates` feature

Signed-off-by: Francesco Romani <fromani@redhat.com>

* e2e: deferred: assertions to avoid unwanted changes

Add assertions to ensure deferred updates don't
affect unrelated nodes.
In the e2e tests we target one of the worker nodes,
so we check control-plane nodes are not affected
by the testing changes.

Checking different MCPs would be equivalent or better,
but this is the simplest way to check because
the worker/control-plane split is always available.

Signed-off-by: Francesco Romani <fromani@redhat.com>

* e2e: deferred: check revert restores node condition

Extract helper code to enable verification of node
state, because we are starting to see the same snippet
(which can't and shouldn't fail) in few places.

Add check to make sure we revert correctly to node state
on profile removal.

Signed-off-by: Francesco Romani <fromani@redhat.com>

* e2e: deferred: check restore after restart

Add test to verify that node state is correctly restored
after a deferred update is applied (requires a restart)
and then removed

Signed-off-by: Francesco Romani <fromani@redhat.com>

* e2e: push down utilities

enable to add another basic rollback test

Signed-off-by: Francesco Romani <fromani@redhat.com>

* e2e: basic: rollback test

add basic rollback test using the same manifests
we do use in deferred tests. This extra test
partially overlap with existing one, and will
help pinpoint deferred update test failures

Signed-off-by: Francesco Romani <fromani@redhat.com>

* makefile: ensure bindata

0_config requires now `pkg/manifests`, so we need
to make sure bindata is generated

Signed-off-by: Francesco Romani <fromani@redhat.com>

* e2e: util: rename ready->wait

rename the "ready" util package to "ready" and change the function
signatures for better readability.

Signed-off-by: Francesco Romani <fromani@redhat.com>

---------

Signed-off-by: Francesco Romani <fromani@redhat.com>